refactor REST and gRPC server
