+++ cmd/shortener/main.go
@@ -7,9 +7,8 @@ import (
 	_ "github.com/shulganew/shear.git/docs" // docs is generated by Swag CLI, you have to import it.
 	"github.com/shulganew/shear.git/internal/app"
 	"github.com/shulganew/shear.git/internal/config"
-	"github.com/shulganew/shear.git/internal/handler/grpc/servgrpc"
+
 	"github.com/shulganew/shear.git/internal/handler/http/router"
-	"github.com/shulganew/shear.git/internal/handler/http/servrest"
 	"github.com/shulganew/shear.git/internal/service"
 
 	"github.com/shulganew/shear.git/pkg/shutdown"
@@ -56,19 +55,19 @@ func main() {
 	componentsErrs := make(chan error, 1)
 
 	// Start web server.
-	webDone := servrest.Shortener(ctx, conf, componentsErrs, rt)
+	restDone := app.StartREST(ctx, conf, componentsErrs, rt)
 
 	// Start gRPC server.
-	grpcDone := servgrpc.Shortener(ctx, short, conf, db, del, componentsErrs)
+	grpcDone := app.StartGRPC(ctx, short, conf, db, del, componentsErrs)
 
 	// Update deleted shorts from common fan-In channel.
-	delDone := app.DeleteShort(ctx, short, delCh)
+	delDone := service.DeleteShort(ctx, short, delCh)
 
 	//Graceful shutdown.
 	shutdown.Graceful(ctx, cancel, componentsErrs)
 
 	// Waiting http server shuting down.
-	<-webDone
+	<-restDone
 
 	// Waiting gRPC server shuting down.
 	<-grpcDone
