+++ internal/handler/grpc/servgrpc/ghandlers/del_test.go
@@ -14,7 +14,6 @@ import (
 	"google.golang.org/grpc/test/bufconn"
 
 	"github.com/golang/mock/gomock"
-	"github.com/shulganew/shear.git/internal/app"
 	"github.com/shulganew/shear.git/internal/config"
 	pb "github.com/shulganew/shear.git/internal/handler/grpc/proto"
 	"github.com/stretchr/testify/assert"
@@ -25,7 +24,6 @@ import (
 )
 
 func TestGRPCDel(t *testing.T) {
-	app.InitLog()
 	// Buffer for gRPC connection emulation.
 	bufSize := 1024 * 1024
 	var lis *bufconn.Listener
@@ -49,7 +47,7 @@ func TestGRPCDel(t *testing.T) {
 	// Create Del service.
 	delCh := make(chan service.DelBatch)
 	del := service.NewDelete(delCh, &configApp)
-	app.DeleteShort(context.Background(), serviceURL, delCh)
+	service.DeleteShort(context.Background(), serviceURL, delCh)
 
 	// Register gRPC server.
 	us := NewUsersServer(serviceURL, &configApp, nil, del)
